backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
neo4j==5.14.1
python-dotenv==1.0.0
bcrypt==4.1.1
PyJWT==2.8.0
Werkzeug==3.0.1